+++ geotrek/common/tests/test_parsers.py
@@ -113,12 +113,12 @@ class AttachmentLegendParser(AttachmentParser):
 
 class ParserTests(TestCase):
     def test_bad_parser_class(self):
-        with self.assertRaisesRegex(CommandError, "Failed to import parser class 'DoesNotExist'"):
+        with self.assertRaisesRegex(Exception, "Failed to import parser class 'geotrek.common.tests.test_parsers.DoesNotExist'"):
             call_command('import', 'geotrek.common.tests.test_parsers.DoesNotExist', '', verbosity=0)
 
     def test_bad_parser_file(self):
-        with self.assertRaisesRegex(CommandError, "Failed to import parser file 'geotrek/common.py'"):
-            call_command('import', 'geotrek.common.DoesNotExist', '', verbosity=0)
+        with self.assertRaisesRegex(Exception, "module 'parsers' has no attribute 'DoesNotExist'"):
+            call_command('import', 'DoesNotExist', '', verbosity=0)
 
     def test_no_filename_no_url(self):
         with self.assertRaisesRegex(CommandError, "File path missing"):